Refactored OwnedCoinDto data types

+++ feature_home/src/main/java/com/ryankoech/krypto/feature_home/data/dto/owned_coin/OwnedCoinDto.kt
@@ -10,8 +10,8 @@ data class OwnedCoinDto(
     @PrimaryKey(autoGenerate = false)
     val id : String,
     val symbol : String,
-    val value : Float,
-    val change : Double,
-    val amount : Float,
+    val value : Double,
+    val change : Float,
+    val amount : Double,
     val icon : String,
 )
